requirements.txt
csaps==1.0.3
cycler==0.10.0
Cython==0.29.22
hdbscan==0.8.27
joblib==1.0.1
kiwisolver==1.3.1
llvmlite==0.36.0
matplotlib==3.3.4
numba==0.53.1
numpy==1.19.5
pandas==1.2.3
Pillow==8.1.2
pybind11==2.6.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
scikit-learn==0.24.1
scipy==1.6.1
Shapely==1.7.1
six==1.15.0
sklearn==0.0
sophuspy==0.0.8
threadpoolctl==2.1.0
